Two fixes: Fix private variable nameing, and create copys on adding lines.
Using private variable names so as not to conflict with the functions by
the same name.

Need to use copy.copy() to make copies when appending/inserting new lines
in the chart file - else will end up with a list of items that all point
to the same memory location.