A Class Representing a Slide - a chart layed out graphically.
This class handles the graphichal layout of the text onto a surface.
It also provides some supporting info.  It exposes a list of page breaks
and default characters with (x,y) coords where they should be used.  Its
main feature is a surface with all the text layed out.

The intention is that this class uses Chart, Line and Section to create
the layout and the the Project will handle the masking, moving and event
loop.