Bug Fix: Fix faulty parsing of default characters.
We would get an index error when trying to use the line.defaultCharacters
array.  The parser would fail to set all the positions when using the
markers in the chart file.

Now we set the default character to 0, then look for markers, and if any
are found we then set the default chr  to that position.