Bug Fix: Mask needs to be same size as viewport and blit onto same position.
The mask was the size of the full monitor. When creating a mask by hand it
would not be visible at first because of the black border. Now it will
start masking imeadately as it only masks the viewport.

Because we only mask the viewport, we must blit onto the same position as
the viewport when displaying - else it looks funny.

Also removed now unneeded code. The position calculation would try to
subtract the viewport and slide widths - but they are now the same and
so why bother.