Code to display the letters visible in the viewport centered in the screen.
We are changing from using a mask on top of the viewport, to croping the
viewport window itself. This allows for better memory useage, but also
simplifys some of the centering and moving calculations.

We now create a background surface the same size as the viewport, and blit
the letters visible in the viewport onto the background. The background is
white or red/green and the letters are black. The background is then blitted
onto the screen at a centered position. The screen has been filled with
black. The net effect is to have the screen act as the black of the mask -
 although there is nothing under the black.