Code to use viewport for line/letter isolation.
Instead of using a seperate mask, we now crop and center the viewport
itself.

Note that resizing the view and centering it both affect the
coordinates, so it's possible to have funny behavior. Also, I've left the
line clearing any mask applied when enter is pressed a fourth time. This
should be removed after all the other masking code is converted to using
the viewport.