Code to move the viewport linewise when only showing a line or letter.
We check to see if a mask is active (showing a single line or letter) and
if so, we jump the view on a per line basis, rather than a per pixel basis.
This makes it easier to navigate when a mask is active.

Also we make the viewport the correct size for the line it is showing - so
we have correct masking as we scroll up and down.