Bug Fix: Fix crash when going from custom mask to line/chr isolation view.
The viewport could end up with a width or height (or both) that were less
than zero. This crashed the program when trying to create a background
surface the size of the viewport.

Now we calculate the amount to inflate the viewport based on its current
size. This way if the view will jump to the correct size even if starting
from a non-default size.