Bug Fix: Fix crash when shrinking viewport manually.
The view could be shrunk to have negative (x,y) coords, which caused a
crash when creating a background surface.

We now check for negative values and, if found, set the coords to 0. This
allows us to blank the screen (create a viewport of zero height or width),
but won't crash the program.