Bug Fix: Remove end of chart page bookmark to prevent blank screens on page down.
Is bad form to show a blank screen to the patient. This would happen every
time the doctor paged down to the end, as the last page marker would
be the coordinates for the end of the chart. However, as page markers
represent the top of the screen to be displayed, a blank screen would be
shown.

Removing the code appending a marker to the end of all charts fixes this.