Bug Fix: Prevent blank screen when moving left/right per character mode.
I was checking that the left side of the view didn't pass the right side
of the slide, and that the right side of the view didn't pass the left side
of the slide. But this meant that the view center had already gone past any
letters, and was showing a blank space.

By checking the view center against the side of the slide, we prevent
showing any blank spaces, but still allow for large letter masks to overlap
the side of the slide if necessary.