Code to check vertical margins - such that no lines are cut off.
This function checks to see if the top line is more than 50% visible. If so,
we use that as the top line, if not, we use the next line down as the top
line. We then calculate the bottom-most line that will fit fully in the
viewport. Then we compute the whitespace needed on either side to center
all the lines in the view. Finally we move the view to the correct
position.