Code: In single line mode, resize viewport to correct size when switching charts.
In single line mode, if the user switches charts, we now check both the
viewport size and vertical centering after the new slide surface is created.
This ensures that we never end up with off center letters or incorrectly
sized masks.

Note that the line sizing only happens in single line mode, but the vertical
centering happens everytime charts are switched. The idea is that the letters
are always centered, no matter how you get to them.