Bug Fix: Rework pageDown function to fix skipping of markers.
It was possible for the checkVerticalCentering function to move the viewport
such that the pageDown function would ignore the next marker. The view
would be relocated such that the current marker was just outside the closest
marker pixel boundary - this would cause pageDown to continiously page to
the same spot.

We now use a different algorithm which prevents this as it doesn't rely on
a hardcoded pixel value to determin the current page marker.